requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
requests
browser-use
openai-agents